pyproject.toml
[project]
name = "hamlet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "cyrilli", email = "948875507@qq.com" }
]
requires-python = ">=3.10"
dependencies = [
  "huggingface-hub>=0.31.2",
  "requests>=2.32.3",
  "rich>=13.9.4",
  "jinja2>=3.1.4",
  "pillow>=10.0.1", # Security fix for CVE-2023-4863: https://pillow.readthedocs.io/en/stable/releasenotes/10.0.1.html
  "python-dotenv",
  "litellm>=1.60.2,<1.70",
  "openai>=1.58.1",
  "datasets==4.1.1",
  "pathvalidate>=3.3.1",
]

[project.scripts]
hamlet = "hamlet:main"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[project.optional-dependencies]
train = [
    "torch>=2.7.0",
    "transformers",
    "accelerate>=1.4.0",
    "requests",
    "peft",
    "wandb",
    "trl>=0.17.0",
    "vllm>=0.9.2",
    "liger-kernel>=0.5.10",
    "deepspeed",
]
serve = [
  "gradio_pdf>=0.0.22",
  "watchdog>=6.0.0",
  "opentelemetry-sdk>=1.33.1",
  "opentelemetry-exporter-otlp>=1.33.1",
  "openinference-instrumentation-smolagents>=0.1.11",
]
tools = [
  "markitdown[all]>=0.1.2",
  "watchdog>=6.0.0",
  "serpapi",
  "langchain==0.3.25",
  "langchain_text_splitters==0.3.8",
  "selenium",
  "PyMuPDF",
  "puremagic",
  "nest_asyncio",
  "crawl4ai>=0.7.0",
  "loguru",
  "certifi",
  "cryptography",
  "pyOpenSSL",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.2",
]
